Movido processTransaction da nano para dentro do index
