+++ external_modules/nano/methods/index.ts
@@ -1,4 +1,3 @@
 export * from './nanoWebSocket'
 export * from './rpc'
-export * from './processTransaction'
 export * from './findMissingTx'
